refactor code to support prometheus and expvar
