+++ geth/node/node.go
@@ -23,6 +23,7 @@ import (
 	whisper "github.com/ethereum/go-ethereum/whisper/whisperv5"
 	"github.com/status-im/status-go/geth/log"
 	"github.com/status-im/status-go/geth/params"
+	whisperMetrics "github.com/status-im/status-go/metrics/whisper"
 )
 
 // node-related errors
@@ -164,7 +165,7 @@ func activateShhService(stack *node.Node, config *params.NodeConfig) error {
 		whisperService := whisper.New(nil)
 
 		// enable metrics
-		whisperService.RegisterEnvelopeTracer(&WhisperEnvelopeTracer{})
+		whisperService.RegisterEnvelopeTracer(&whisperMetrics.EnvelopeTracer{})
 
 		// enable mail service
 		if whisperConfig.EnableMailServer {
