+++ metrics/metrics_prometheus.go
@@ -1,0 +1,16 @@
+// +build prometheus
+
+package metrics
+
+import (
+	"log"
+	"net/http"
+
+	"github.com/prometheus/client_golang/prometheus/promhttp"
+)
+
+// StartMetricsServer starts a new HTTP server with prometheus handler.
+func StartMetricsServer(addr string) {
+	http.Handle("/metrics", promhttp.Handler())
+	log.Fatal(http.ListenAndServe(addr, nil))
+}
